requirements.txt
dj-database-url==0.5.0
Django==2.2.6
django-cors-headers==3.1.1
djangorestframework==3.10.3
gunicorn==19.9.0
joblib==0.14.0
numpy==1.17.2
pandas==0.25.1
python-dateutil==2.8.0
pytz==2019.2
scikit-learn==0.21.3
scipy==1.3.1
six==1.12.0
sklearn==0.0
sqlparse==0.3.0
whitenoise==4.1.4
psycopg2==2.7.1
